Added scenes.py
